+++ src/ImplicitSphere.cpp
@@ -8,41 +8,41 @@ Sphere::Sphere(FieldFunction f) :
 	compute_bounds();
 }
 
-Sphere::Sphere(FieldFunction f, float iso) :
+Sphere::Sphere(FieldFunction f, double iso) :
 	Primitive(f, iso)
 {
 	compute_bounds();
 }
 
-Sphere::Sphere(FieldFunction f, float iso, float radius) :
+Sphere::Sphere(FieldFunction f, double iso, double radius) :
 	Primitive(f, iso, radius)
 {
 	compute_bounds();
 }
 
-float Sphere::Evaluate(const glm::vec3& p)
+double Sphere::Evaluate(const glm::dvec3& p)
 {
 	return Primitive::Evaluate(getDistance(p));
 }
 
-float Sphere::FieldValue(const glm::vec3& p)
+double Sphere::FieldValue(const glm::dvec3& p)
 {
 	return Primitive::FieldValue(getDistance(p));
 }
 
-glm::vec3 Sphere::GetStartVertex()
+glm::dvec3 Sphere::GetStartVertex()
 {
-	return project(glm::vec3(0, 0, 0));
+	return project(glm::dvec3(0, 0, 0));
 }
 
-glm::vec3 Sphere::GetCenterVertex()
+glm::dvec3 Sphere::GetCenterVertex()
 {
 	return glm::vec3(0, 0, 0);
 }
 
 void Sphere::compute_bounds()
 {
-	std::list<glm::vec3> points;
+	std::list<glm::dvec3> points;
 	points.push_back(glm::vec3(m_radius, 0, 0) );
 	points.push_back(glm::vec3(-m_radius, 0, 0));
 	points.push_back(glm::vec3(0, m_radius, 0) );
@@ -53,13 +53,13 @@ void Sphere::compute_bounds()
 }
 
 
-glm::vec3 Sphere::Normal(const glm::vec3& point)
+glm::dvec3 Sphere::Normal(const glm::dvec3& point)
 {
-	if (point == glm::vec3(0, 0, 0)) return glm::vec3(0, 1, 0);
+	if (point == glm::dvec3(0, 0, 0)) return glm::dvec3(0, 1, 0);
 	return glm::normalize(point);
 }
 
-float Sphere::getDistance(const glm::vec3& p)
+double Sphere::getDistance(const glm::dvec3& p)
 {
 	return glm::length(p);
 }
